Reworked CSS for Dashboard, QuizEdit, Header, QuizCard, PlayQuestionCard

+++ frontend/src/components/QuizCard.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
-import { Button, Card, Image, Modal, Badge } from 'react-bootstrap';
+import { Button, Card, Image, Modal, Badge, Stack } from 'react-bootstrap';
 import API from '../utils/API';
 import { useAuth } from '../utils/Auth';
 import { Link, useNavigate } from 'react-router-dom';
@@ -8,6 +8,16 @@ import useQuizFetch from '../hooks/useQuizFetch';
 import useAdminStatus from '../hooks/useAdminStatus';
 import Loading from './Loading';
 import CountdownTimer from './CountdownTimer';
+import { BsQuestionCircle, BsStopwatch, BsCalendarWeek, BsPeople, BsCardList } from 'react-icons/bs';
+import styled from 'styled-components';
+
+const BadgeLink = styled(Link)`
+  text-decoration: none;
+  color: white;
+  &:hover {
+    color: lightgray;
+  }
+`;
 
 const QuizCard = ({ empty, quizid, fetchAllQuizzes }) => {
   const { token } = useAuth();
@@ -16,6 +26,7 @@ const QuizCard = ({ empty, quizid, fetchAllQuizzes }) => {
   const { quiz, quizLoading, fetchQuiz } = useQuizFetch(token, quizid);
   const { adminStatus, fetchAdminStatus } = useAdminStatus(token, sessionId);
   const [loading, setLoading] = useState(false);
+  const [quizStarted, setQuizStarted] = useState(false);
   const [questions, setQuestions] = useState([]);
   const [duration, setDuration] = useState(0);
   const [showStartPopup, setShowStartPopup] = useState(false);
@@ -33,10 +44,21 @@ const QuizCard = ({ empty, quizid, fetchAllQuizzes }) => {
     );
   }
 
-  // Set sessionid
-  useEffect(() => {
-    if (quiz.active) setSessionId(quiz.active);
-  }, [quiz]);
+  // Set sessionid and timer to check number of players
+  useEffect(async () => {
+    const intervalFetch = async (token, sid) => await fetchAdminStatus(token, sid);
+
+    if (quiz.active) {
+      setSessionId(quiz.active);
+    }
+    let interval;
+    if (quiz.active && adminStatus.position === -1 && !quizStarted) {
+      interval = setInterval(() => intervalFetch(token, quiz.active), 1000);
+    } else {
+      clearInterval(interval);
+    }
+    return () => clearInterval(interval);
+  }, [quiz, quizStarted, adminStatus]);
 
   // Fetch questions from backend
   useEffect(async () => {
@@ -68,13 +90,15 @@ const QuizCard = ({ empty, quizid, fetchAllQuizzes }) => {
   // Delete Button Handler
   const handleDelete = async (event) => {
     event.preventDefault();
-    setLoading(true);
-    const data = await API.deleteQuiz(token, quizid);
-    setLoading(false);
-    if (data.error) {
-      console.log(data.error);
-    } else {
-      fetchAllQuizzes(token);
+    if (confirm(`Are you sure you want to delete ${quiz.name}?`)) {
+      setLoading(true);
+      const data = await API.deleteQuiz(token, quizid);
+      setLoading(false);
+      if (data.error) {
+        console.log(data.error);
+      } else {
+        fetchAllQuizzes(token);
+      }
     }
   }
 
@@ -105,6 +129,7 @@ const QuizCard = ({ empty, quizid, fetchAllQuizzes }) => {
       } else {
         fetchQuiz(token, quizid);
         fetchAdminStatus(token, sessionId);
+        setQuizStarted(true);
         setQuestionInProgress(true);
       }
       setLoading(false);
@@ -120,6 +145,7 @@ const QuizCard = ({ empty, quizid, fetchAllQuizzes }) => {
       console.error(data.error);
     } else {
       fetchQuiz(token, quizid);
+      setQuizStarted(false);
       setShowStopPopup(true);
       setQuestionInProgress(false);
     }
@@ -137,48 +163,74 @@ const QuizCard = ({ empty, quizid, fetchAllQuizzes }) => {
     return window.location.href.replace('dashboard', `quiz/join/${sessionid}`);
   }
 
+  // Format duration to string
+  const formatDuration = (duration) => {
+    const m = Math.floor(duration / 60);
+    const s = duration % 60;
+    if (m && s) return `${m}:${String(s).padStart(2, '0')}`;
+    if (m) return `${m}:00`;
+    if (s) return `0:${String(s).padStart(2, '0')}`
+  }
+
   if (loading || quizLoading) return <Loading/>;
 
   return (
     <>
       <Card className='shadow'>
         <Card.Body>
+          {/* Title / Active Row */}
           <div className='d-flex align-items-center justify-content-between mb-1'>
             <h4 className='mb-0'>{quiz.name}</h4>
-            <div>
-              <Badge pill>{ questions.length } question{ questions.length === 1 ? '' : 's' }</Badge>
-              <Badge pill bg='secondary' className='ms-2'>{duration} seconds</Badge>
+              { quiz.active
+                ? <Badge className='ms-auto' pill bg='success'><BadgeLink to={`/quiz/join/${quiz.active}`} target='_blank'>Active: {quiz.active}</BadgeLink></Badge>
+                : <Badge className='ms-auto' pill bg='secondary'>{'Inactive'}</Badge>
+              }
+          </div>
+          {/* Content Cell */}
+          <div className='d-flex flex-column gap-2'>
+            { quiz.active && adminStatus.position !== -1 &&
+              <CountdownTimer
+                timer={{
+                  start: adminStatus.isoTimeLastQuestionStarted,
+                  duration: adminStatus.questions[adminStatus.position].duration
+                }}
+                onEnd={() => setQuestionInProgress(false)}
+              />
+            }
+            <div className='d-flex justify-content-between'>
+              {/* Left Column */}
+              <Stack gap={2} className=''>
+                <div><BsQuestionCircle/> { questions.length } question{ questions.length === 1 ? '' : 's' }</div>
+                <div><BsStopwatch/> {formatDuration(duration)}</div>
+                <div><BsCalendarWeek/> {new Date(quiz.createdAt).toLocaleString()}</div>
+                { quiz.thumbnail &&
+                  <div>
+                    <Image fluid thumbnail src={quiz.thumbnail} alt='No image' width='180px' height='100%'/>
+                  </div>
+                }
+              </Stack>
+              {/* Right Column */}
+              <div className='d-flex align-items-center justify-content-between p-3 bg-light border border-2 border-secondary rounded'>
+                {/* Active Game Column */}
+                { quiz.active &&
+                  <Stack gap={2} className='d-flex align-items-start justify-content-center me-4'>
+                    <div><BsPeople/> {adminStatus.players.length}</div>
+                    <div><BsCardList/> {adminStatus.position === -1 ? 'Lobby' : `Question ${adminStatus.position + 1} of ${adminStatus.questions.length}`}</div>
+                  </Stack>
+                }
+                {/* Buttons Column */}
+                <div className='d-flex align-items-end justify-content-center'>
+                  <Stack gap={2}>
+                    { !quiz.active && <Button variant='primary' onClick={handleEdit}>Edit</Button> }
+                    { !quiz.active && <Button variant='danger' onClick={handleDelete}>Delete</Button> }
+                    { !quiz.active && <Button variant='success' onClick={handleStart}>Start</Button> }
+                    { quiz.active && <Button variant='warning' onClick={handleAdvance} disabled={questionInProgress || !adminStatus.players.length}>Next</Button> }
+                    { quiz.active && <Button variant='danger' onClick={handleStop}>Stop</Button> }
+                  </Stack>
+                </div>
+              </div>
             </div>
           </div>
-          <p>Created: {new Date(quiz.createdAt).toLocaleString()}</p>
-          <Image thumbnail src={quiz.thumbnail} alt='No image' width='100px' height='100px'/>
-          <p>Owner: {quiz.owner}</p>
-          { quiz.active && <p>Active Room: <Link to={`/quiz/join/${quiz.active}`} target='_blank'>{quiz.active}</Link></p> }
-          { quiz.active && <p>{adminStatus.players.length} player{adminStatus.players.length === 1 ? '' : 's'}</p> }
-          { quiz.active && <p>Current Status: {adminStatus.position === -1 ? 'Lobby' : `Question ${adminStatus.position + 1} of ${adminStatus.questions.length}`}</p> }
-          { quiz.active && adminStatus.position !== -1 &&
-            <CountdownTimer
-              timer={{
-                start: adminStatus.isoTimeLastQuestionStarted,
-                duration: adminStatus.questions[adminStatus.position].duration
-              }}
-              onEnd={() => setQuestionInProgress(false)}
-            />
-          }
-          { !quiz.active && <Button variant='success' onClick={handleStart}>Start</Button> }
-          { quiz.active &&
-            <>
-              <Button variant='warning' onClick={handleAdvance} disabled={questionInProgress}>Advance</Button>
-              <Button variant='danger' onClick={handleStop}>Stop</Button>
-            </>
-          }
-          { !quiz.active &&
-            <>
-              <br/>
-              <Button variant='primary' onClick={handleEdit}>Edit</Button>
-              <Button variant='danger' onClick={handleDelete}>Delete</Button>
-            </>
-          }
         </Card.Body>
       </Card>
       <Modal show={showStartPopup} onHide={() => setShowStartPopup(false)}>
